+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- A couple of very dedicated methods to QueueTaskRepository for retrieving queue tasks by
+  public id, status, queue name and payload content.
+  
+### Changed
+- **[DATABASE MIGRATION NEEDED]** Add unique db index on `queue_task.id_public`.
 
 ## [4.0.1] - 2017-08-21
 ### Changed